CPU命令実行部分を別ファイルに分離
